+++ src/Libpfs/manip/cut.h
@@ -24,6 +24,7 @@
 #ifndef PFS_CUT_H
 #define PFS_CUT_H
 
+#include <cstddef>
 #include <Libpfs/array2d_fwd.h>
 
 //! \brief Cut a rectangle out of images in PFS stream
@@ -35,11 +36,11 @@ namespace pfs
 class Frame;
 
 Frame *cut(const Frame *inFrame,
-           int x_ul, int y_ul, int x_br, int y_br);
+           size_t x_ul, size_t y_ul, size_t x_br, size_t y_br);
 
 template <typename Type>
 void cut(const Array2D<Type> *from, Array2D<Type> *to,
-         int x_ul, int y_ul, int x_br, int y_br);
+         size_t x_ul, size_t y_ul, size_t x_br, size_t y_br);
 }
 
 #include "cut.hxx"
